Remove redundant static cast

+++ src/engine/core/systems/renderer/renderer.cpp
@@ -260,7 +260,7 @@ namespace hob {
             const float pen_y = origin.y;
 
             for (char c : text) {
-                const uint32_t cp = static_cast<uint32_t>(static_cast<unsigned char>(c));
+                const uint32_t cp = static_cast<unsigned char>(c);
                 const Glyph* g = m_debug_font.get_glyph(cp);
                 if (!g) {
                     continue;
